Ajustado Controller, interface/contrato e service para uso do novo padrão de DTO.

+++ src/main/java/com/dutra/empreende_sc/service/EmpreendimentoServiceImpl.java
@@ -2,12 +2,14 @@ package com.dutra.empreende_sc.service;
 
 import com.dutra.empreende_sc.dtos.EmpreendimentoDtoInput;
 import com.dutra.empreende_sc.dtos.EmpreendimentoDtoOutput;
+import com.dutra.empreende_sc.dtos.EmpreendimentoUpdateInput;
 import com.dutra.empreende_sc.entities.Empreendimento;
 import com.dutra.empreende_sc.exceptions.EntidadeNaoEncontradaException;
 import com.dutra.empreende_sc.repository.EmpreendimentoRepository;
 import com.dutra.empreende_sc.service.interfaces.EmpreendimentoService;
 
 import com.dutra.empreende_sc.service.utils.EmpreendimentoMapper;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -28,42 +30,74 @@ public class EmpreendimentoServiceImpl implements EmpreendimentoService {
     @Override
     public EmpreendimentoDtoOutput criar(EmpreendimentoDtoInput empreendimento) {
         Empreendimento novoEmpreendimento = mapper.toEntity(empreendimento);
-
         return mapper.toOutput(repository.save(novoEmpreendimento));
     }
 
     @Transactional(readOnly = true)
     @Override
     public List<EmpreendimentoDtoOutput> listarTodos() {
-        return repository.findAll();
+        return repository.findAll().stream().map(mapper::toOutput).toList();
     }
 
     @Transactional(readOnly = true)
     @Override
     public EmpreendimentoDtoOutput buscarPorId(Long id) {
-        return repository.findById(id).orElseThrow(
-                () -> new EntidadeNaoEncontradaException("Empreendimento não encontrado com o id: " + id)
-        );
+        return mapper.toOutput(repository.findById(id).orElseThrow(
+                () -> new EntidadeNaoEncontradaException("Empreendimento não encontrado para o id: " + id)
+        ));
     }
 
+    @Transactional
     @Override
-    public EmpreendimentoDtoOutput atualizar(Long id, EmpreendimentoDtoInput empreendimentoAtualizado) {
+    public EmpreendimentoDtoOutput atualizar(Long id, EmpreendimentoUpdateInput empreendimentoAtualizado) {
 
-        return repository.findById(id).map(empreendimentoExistente -> {
-            empreendimentoExistente.setNomeEmpreendimento(empreendimentoAtualizado.nomeEmpreendimento());
-            empreendimentoExistente.setNomeEmpreendedor(empreendimentoAtualizado.nomeEmpreendedor());
-            empreendimentoExistente.setMunicipio(empreendimentoAtualizado.municipio());
-            empreendimentoExistente.setSegmento(empreendimentoAtualizado.segmento());
-            empreendimentoExistente.setContato(empreendimentoAtualizado.contato());
-            empreendimentoExistente.setStatus(empreendimentoAtualizado.status());
-            return repository.save(empreendimentoExistente);
-        })
-                .orElseThrow(() -> new RuntimeException("Empreendimento não encontrado com o id: " + id));
+        Empreendimento entidade = repository.findById(id)
+                .orElseThrow(() -> new EntidadeNaoEncontradaException("Empreendimento não encontrado com o id: " + id));
+
+        updateEntityFromDto(empreendimentoAtualizado, entidade);
+
+        Empreendimento entidadeSalva = repository.save(entidade);
+
+        return mapper.toOutput(entidadeSalva);
     }
 
     @Transactional
     @Override
     public void deletar(Long id) {
-        repository.deleteById(id);
+
+        if (!repository.existsById(id)) {
+            throw new EntidadeNaoEncontradaException("Empreendimento não encontrado com o id: " + id);
+        }
+
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException("Referential integrity violation.");
+        }
+    }
+
+    private void updateEntityFromDto(EmpreendimentoUpdateInput dto, Empreendimento entidade) {
+        if (dto == null || entidade == null) {
+            return;
+        }
+
+        if (dto.nomeEmpreendimento() != null) {
+            entidade.setNomeEmpreendimento(dto.nomeEmpreendimento());
+        }
+        if (dto.nomeEmpreendedor() != null) {
+            entidade.setNomeEmpreendedor(dto.nomeEmpreendedor());
+        }
+        if (dto.municipio() != null) {
+            entidade.setMunicipio(dto.municipio());
+        }
+        if (dto.segmento() != null) {
+            entidade.setSegmento(dto.segmento());
+        }
+        if (dto.contato() != null) {
+            entidade.setContato(dto.contato());
+        }
+        if (dto.status() != null) {
+            entidade.setStatus(dto.status());
+        }
     }
 }
